Move clone-find-marked commands into leoFind.py. Delete one file
